ue_project/Scripts/scene_builder.py
"""
UE5 Scene Builder
Automates creation of virtual news studio environments
"""

try:
    import unreal
except ImportError:
    print("Warning: This script must be run within Unreal Engine Python environment")
    unreal = None


class NewsStudioBuilder:
    """Automated news studio scene builder"""
    
    def __init__(self):
        if unreal:
            self.editor_util = unreal.EditorLevelLibrary()
            self.editor_asset_lib = unreal.EditorAssetLibrary()
    
    def create_basic_studio(self, studio_name: str = "NewsStudio_Default"):
        """
        Create a basic news studio setup
        
        Args:
            studio_name: Name for the studio level
        """
        if not unreal:
            print("Must run in UE5 environment")
            return
        
        print(f"[SceneBuilder] Creating studio: {studio_name}")
        
        # Create new level
        # level_path = f"/Game/Maps/{studio_name}"
        # self.editor_util.new_level(level_path)
        
        # Add basic components
        self._add_floor()
        self._add_desk()
        self._add_lighting()
        self._add_cameras()
        
        print(f"[SceneBuilder] ✓ Studio created")
    
    def _add_floor(self):
        """Add floor plane"""
        print("[SceneBuilder] Adding floor...")
        # TODO: Spawn floor actor
        pass
    
    def _add_desk(self):
        """Add news desk"""
        print("[SceneBuilder] Adding news desk...")
        # TODO: Spawn desk actor
        pass
    
    def _add_lighting(self):
        """Setup studio lighting"""
        print("[SceneBuilder] Setting up lighting...")
        # TODO: Add key light, fill light, rim light
        pass
    
    def _add_cameras(self):
        """Add camera actors"""
        print("[SceneBuilder] Adding cameras...")
        # TODO: Add camera actors with different angles
        # - Wide shot
        # - Medium shot
        # - Close-up
        pass
    
    def setup_metahuman(self, metahuman_name: str, position: tuple = (0, 0, 100)):
        """
        Place MetaHuman in scene
        
        Args:
            metahuman_name: Name of MetaHuman asset
            position: (X, Y, Z) position
        """
        print(f"[SceneBuilder] Placing MetaHuman: {metahuman_name}")
        # TODO: Spawn MetaHuman actor at position
        pass


# Utility functions for external calls
def create_news_studio(studio_name: str = "NewsStudio_Default"):
    """Create a news studio (callable from external Python)"""
    builder = NewsStudioBuilder()
    builder.create_basic_studio(studio_name)


if __name__ == "__main__":
    # Example usage when run in UE5
    create_news_studio()
